# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "JobTrackr API",
        "description": "API documentation for JobTrackr - A job application tracking platform",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local Development Server"
        }
    ],
    "paths": {
...
        "/api/profile": {
            "get": {
                "tags": [
                    "Profile"
                ],
                "summary": "Get user profile",
                "operationId": "f1784685f8270527da017b3616288443",
                "responses": {
                    "200": {
                        "description": "User profile retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "user": {
                                            "$ref": "#/components/schemas/User"
                                        },
                                        "unread_notifications_count": {
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Profile"
                ],
                "summary": "Update user profile",
                "operationId": "875e8801180287e4193f1aced9023960",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "John Doe"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "john@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "format": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Profile updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "user": {
                                            "$ref": "#/components/schemas/User"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/stats/overview": {
            "get": {
                "tags": [
                    "Statistics"
                ],
                "summary": "Get dashboard overview statistics",
                "operationId": "9b8d735ee30cb7e538ea6b040874bd8e",
                "responses": {
                    "200": {
                        "description": "Dashboard statistics"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/stats/monthly": {
            "get": {
                "tags": [
                    "Statistics"
                ],
                "summary": "Get applications per month",
                "operationId": "2813379db993b8e2f342372a6199235b",
                "parameters": [
                    {
                        "name": "year",
                        "in": "query",
                        "description": "Filter by year",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 2025
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Monthly statistics"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/stats/by-status": {
            "get": {
                "tags": [
                    "Statistics"
                ],
                "summary": "Get detailed statistics by status",
                "operationId": "1ff062dbda4b26f32480e9ab5cdae241",
                "responses": {
                    "200": {
                        "description": "Status breakdown"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/stats/top-companies": {
            "get": {
                "tags": [
                    "Statistics"
                ],
                "summary": "Get most targeted companies",
                "operationId": "b93763220f054d3c2dc5bd43e20b4834",
                "parameters": [
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Number of companies to return",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Top companies list"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/stats/recent-activity": {
            "get": {
                "tags": [
                    "Statistics"
                ],
                "summary": "Get recent application activity",
                "operationId": "ae902a538d0cbda9c5f68dfaee0543fd",
                "parameters": [
                    {
                        "name": "days",
                        "in": "query",
                        "description": "Number of days to look back",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 30
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Recent activity"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "User": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "John Doe"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "john@example.com"
                    },
                    "avatar": {
                        "type": "string",
                        "example": "avatars/avatar.jpg",
                        "nullable": true
                    },
                    "cv_path": {
                        "type": "string",
                        "example": "cvs/resume.pdf",
                        "nullable": true
                    },
                    "email_verified_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}